Add league subscriptions create endpoint
